backend/buildingExtractor/ml-model/requirements_py313.txt
# Core dependencies for aerial imagery annotation and building detection
# Compatible with Python 3.13
opencv-python>=4.8.0
numpy>=1.24.0
matplotlib>=3.7.0
pillow>=10.0.0
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
folium>=0.14.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
albumentations>=1.3.0
tqdm>=4.65.0
pandas>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0
geopy>=2.3.0
plotly>=5.15.0
dash>=2.11.0
dash-bootstrap-components>=1.4.0
jupyter>=1.0.0
ipywidgets>=8.0.0
openpyxl>=3.1.0
flask>=2.3.0
werkzeug>=2.3.0
